Joined CalendarController with GoogleAuthController, rename to GoogleController
